refactor(cockpit_ui): extract settings-data gather helpers (refs #1354)
Move the settings-screen data layer — ``_gather_settings_data`` plus its
15 supporting helpers (role rows, status dots, budget formatting,
``du -sk`` disk-size shell-out, recent-tasks-by-account collector, etc.) —
into a new ``pollypm.cockpit_settings_gather`` module. ``cockpit_ui``
re-exports the names so existing imports and
``monkeypatch.setattr("pollypm.cockpit_ui._collect_recent_tasks_by_account", ...)``
test hooks keep working.

This is the first slice of the rolling ``cockpit_ui.py`` god-module split
tracked by #1354. Pure module-boundary move: no behavior changes, no
public-API churn beyond test patches that target the moved symbols.
``cockpit_ui.py`` shrinks 18,368 → 17,807 LOC (-561); the new module is
~580 LOC and ruff-clean. Remaining cockpit_ui.py ruff errors dropped
from 53 → 50 (the moved code's transitive imports are gone).

- src/pollypm/cockpit_settings_gather.py: new module with role helpers,
  settings-dot/disk/humanize formatters, budget/usage helpers,
  recent-task collector, session-refs grouping, and
  ``_gather_settings_data`` itself.
- src/pollypm/cockpit_ui.py: drop the moved block; replace with a
  single re-export ``from pollypm.cockpit_settings_gather import (…)``.
  Trim now-unused upstream imports (``advisories_for``, ``resolve_alias``,
  ``resolve_role_assignment``, ``load_cached_account_usage``,
  ``history_rationale_for_account``, ``history_rationale_for_project``,
  ``load_settings_history``, ``collect_settings_projects``).
- tests/test_settings_ui.py: redirect five monkeypatches that target
  ``_gather_settings_data``'s dependency names from ``pollypm.cockpit_ui``
  to the new ``pollypm.cockpit_settings_gather`` module so the test
  seams continue to override the same lookups.
